pyproject.toml
[project]
description = "A simple ecies demo based on Litestar"
name = "eciespy-demo"
version = "0.1.0"

readme = "README.md"
requires-python = ">=3.13"

dependencies = [
  "eciespy>=0.4.6",
  "litestar>=2.16.0",
  "msgspec>=0.19.0",
  "uvicorn>=0.35.0",
  "uvloop>=0.21.0; sys_platform != 'win32'",
]

[tool.uv]
dev-dependencies = [
  "mypy>=1.17.0",
  "pytest>=8.4.1",
  "pytest-asyncio>=1.1.0",
  "ruff>=0.12.4",
]

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "session"
asyncio_mode = "auto"
